add current thread parameter to sched_unblock
The previous code did not pass the current thread
to sched_unblock, even though the CFS scheduler requires
access to the current thread to update vruntime before
making a preemption decision.

We now pass the current thread, but we also ensure that
the current ready queue is locked, allowing the scheduler
safe access to the metadata in the ready queue.